moved problem in assembler
