+++ src/polyfem/main.cpp
@@ -8,9 +8,11 @@
 #include <polyfem/utils/StringUtils.hpp>
 #include <polyfem/utils/Logger.hpp>
 
-#include <polyfem/problem/Problem.hpp>
+#include <polyfem/assembler/Problem.hpp>
 #include <polyfem/time_integrator/ImplicitTimeIntegrator.hpp>
 #include <polyfem/assembler/AssemblerUtils.hpp>
+#include <polyfem/problem/ProblemFactory.hpp>
+
 #include <polyfem/utils/JSONUtils.hpp>
 
 #include <geogram/basic/command_line.h>
