+++ src/polyfem/mesh/Obstacle.cpp
@@ -10,7 +10,7 @@
 
 namespace polyfem
 {
-	using namespace problem;
+	using namespace assembler;
 	using namespace utils;
 
 	namespace mesh
