+++ src/polyfem/solver/CMakeLists.txt
@@ -23,4 +23,4 @@ target_sources(polyfem PRIVATE ${SOURCES})
 ################################################################################
 
 add_subdirectory(line_search)
-add_subdirectory(forms)
+# add_subdirectory(forms)
