+++ src/polyfem/state/StateInit.cpp
@@ -1,5 +1,8 @@
 #include <polyfem/State.hpp>
 
+#include <polyfem/problem/ProblemFactory.hpp>
+#include <polyfem/assembler/GenericProblem.hpp>
+
 #include <polyfem/autogen/auto_p_bases.hpp>
 #include <polyfem/autogen/auto_q_bases.hpp>
 
@@ -363,9 +366,9 @@ namespace polyfem
 		if (!args.contains("preset_problem"))
 		{
 			if (assembler.is_scalar(formulation()))
-				problem = ProblemFactory::factory().get_problem("GenericScalar");
+				problem = std::make_shared<assembler::GenericScalarProblem>("GenericScalar");
 			else
-				problem = ProblemFactory::factory().get_problem("GenericTensor");
+				problem = std::make_shared<assembler::GenericTensorProblem>("GenericTensor");
 
 			problem->clear();
 			if (!args["time"].is_null())
